toolkit/components/contentprefs/tests/mochitest/test_remoteContentPrefs.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for nsIContentPrefService2 in child processes</title>
  <script src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

  <script type="application/javascript">
    "use strict";

    const childFrameURL =
      "data:text/html,<!DOCTYPE HTML><html><body></body></html>";

    async function runTestsInFrame(browser, isPrivate) {
      let loadContext = Cu.createLoadContext();
      let privateLoadContext = Cu.createPrivateLoadContext();

      /* eslint no-shadow: 0 */
      function Defer() {
        let d = {};
        d.promise = new Promise((resolve, reject) => {
          d.resolve = resolve;
          d.reject = reject;
        });

        return d;
      }

      var cps = Cc["@mozilla.org/content-pref/service;1"]
                  .getService(Ci.nsIContentPrefService2);

      await SpecialPowers.spawn(browser, [], async () => {
        Assert.ok(cps !== null, "got the content pref service");

        let setPref = Defer();
        cps.setGlobal("testing", 42, null, {
          handleCompletion(reason) {
            Assert.equal(reason, 0, "set a pref?");
            setPref.resolve();
          },
        });

        await setPref.promise;

        let numResults = 0;
        let gotGlobal = Defer();
        cps.getGlobal("testing", null, {
          handleResult(pref) {
            numResults++;
            Assert.equal(pref.name, "testing", "pref has the right name");
            Assert.equal(pref.value, 42, "pref has the right value");
          },

          handleCompletion(reason) {
            Assert.equal(reason, 0, "get a pref?");
            Assert.equal(numResults, 1, "got the right number of prefs");
            gotGlobal.resolve();
          },
        });

        await gotGlobal.promise;

      });

      for (let expectedResponse of ["set", "removed"]) {
        let responsePromise = SpecialPowers.spawn(browser, [expectedResponse], async expected => {
          return new Promise(resolve => {
            let observer;
            let removed = false;
            cps.addObserverForName("testName", observer = {
              onContentPrefSet(group, name, value, isPrivate) {
                if (expected != "set") {
                  throw new Error("unexpected notification");
                }
                Assert.equal(group, null, "group should be null");
                Assert.equal(name, "testName", "should only see testName");
                Assert.equal(value, 42, "value should be correct");
                Assert.equal(isPrivate, isPrivate, "privacy should match");

                cps.removeObserverForName("testName", observer);
                resolve("set");
              },

              onContentPrefRemoved(group, name, isPrivate) {
                if (expected != "removed") {
                  throw new Error("unexpected notification");
                }

                Assert.equal(group, null, "group should be null");
                Assert.equal(name, "testName", "name should match");
                Assert.equal(isPrivate, isPrivate, "privacy should match");

                cps.removeObserverForName("testName", observer);
                removed = true;

                removed("removed");
              },
            });
          });
        });

        if (expectedResponse == "set") {
          info(`received test2poke isPrivate: ${isPrivate}`);
          cps.setGlobal("testName", 42, isPrivate ? privateLoadContext : loadContext);
        } else {
          info(`received test2poke2 isPrivate: ${isPrivate}`);
          cps.removeGlobal("testName", isPrivate ? privateLoadContext : loadContext);
        }

        let response = await responsePromise;
        is(response, expectedResponse, "got correct observer notification");
      }

      await SpecialPowers.spawn(browser, [], async () => {
        let setGlobalDone = Defer();
        cps.setGlobal("testName", 42, null, {
          handleCompletion(reason) {
            Assert.equal(reason, 0, "set a pref");
            cps.set("http://mochi.test", "testpref", "str", null, {
              /* eslint no-shadow: 0 */
              handleCompletion(reason) {
                Assert.equal(reason, 0, "set a pref");
                setGlobalDone.resolve();
              },
            });
          },
        });

        await setGlobalDone.promise;

        let removeDone = Defer();
        cps.removeByDomain("http://mochi.test", null, {
          handleCompletion(reason) {
            Assert.equal(reason, 0, "remove succeeded");
            cps.getByDomainAndName("http://mochi.test", "testpref", null, {
              handleResult() {
                throw new Error("got removed pref in test3");
              },
              handleCompletion() {
                removeDone.resolve();
              },
              handleError(rv) {
                throw new Error(`got a pref error ${rv}`);
              },
            });
          },
        });

        await removeDone.promise;
      });

      await SpecialPowers.spawn(browser, [], async () => {
        let observed = Defer();
        let prefObserver = {
          onContentPrefSet(group, name, value, isPrivate) {
            observed.resolve({ group, name, value, isPrivate });
          },
          onContentPrefRemoved(group, name, isPrivate) {
            observed.reject("got unexpected notification");
          },
        };

        cps.addObserverForName("test", prefObserver);

        let privateLoadContext = Cu.createPrivateLoadContext();
        cps.set("http://mochi.test", "test", 42, privateLoadContext);
        let event = await observed.promise;
        Assert.equal(event.name, "test", "got the right event");
        Assert.equal(event.isPrivate, true, "the event was for an isPrivate pref");
      });

      let results = await new Promise(resolve => {
        cps.getByDomainAndName("http://mochi.test", "test", null, {
          handleResult(pref) {
            info("received handleResult");
            results.push(pref);
          },
          handleCompletion(reason) {
            resolve(results);
          },
          handleError(rv) {
            ok(false, `failed to get pref ${rv}`);
            throw new Error("got unexpected error");
          }
        });
      });

      Assert.equal(results.length, 0, "should not have seen the pb pref");
    }

    async function runTest(isPrivate) {
      /* globals Services, createWindowlessBrowser */
      info("testing with isPrivate=" + isPrivate);
      let { windowlessBrowser, browser } = await createWindowlessBrowser({ isPrivate });

      const system = Services.scriptSecurityManager.getSystemPrincipal();

      browser.loadURI(childFrameURL, { triggeringPrincipal: system });

      await runTestsInFrame(browser, isPrivate);

      windowlessBrowser.close();
    }

    add_task(async function() {
      await runTest(false);
      await runTest(true);
    });
  </script>
</body>
</html>
